[Version 0.0.2g] Fixes
~ Fixed multiple projectile Abilities
~ Fixed Targeted Projectiles
~ Changed xpos and ypos to float
~ Fixed Stacking Effects

+++ src/Game/EnemyMeleeBeserker.java
@@ -42,7 +42,7 @@ public class EnemyMeleeBeserker extends EnemyMelee {
 	}
 
 	void attack(int i) {
-		ability[i].useAbility(direction, xpos, ypos, width);
+		ability[i].useAbility(direction, (int) xpos, (int) ypos, width);
 	}
 
 	void move() {
